--- pages/dashboard/dashboard.ejs
@@ -1,14 +1,0 @@
-Hi.
-
-<% var cols = ['short_name', 'title', 'start_date', 'end_date', 'mode']; %>
-
-<table border=1>
-    <% assessments.forEach(function(a) { %>
-    <tr>
-    <% cols.forEach(function(cell) { %>
-    <td><%= a[cell] %></td>
-    <% }); %>
-    <td><%= JSON.stringify(a); %></td>
-</tr>
-    <% }); %>
-</table>
