+++ pages/partials/navbarPlain.ejs
@@ -1,4 +1,5 @@
 
       <% if (is_administrator) { %>
       <li class="nav-item <% if (navPage == "admin") { %>active<% } %>"><a class="nav-link" href="<%= plainUrlPrefix %>/administrator/overview">Admin</a></li>
+      <li class="nav-item <% if (navPage == "adminDashboard") { %>active<% } %>"><a class="nav-link" href="<%= plainUrlPrefix %>/administrator/dashboard">Dashboard</a></li>
       <% } %>
