+++ server.js
@@ -791,7 +791,7 @@ app.use('/pl/course/:course_id/question/:question_id/text', [
 
 app.use('/pl/administrator', require('./middlewares/authzIsAdministrator'));
 app.use('/pl/administrator/overview', require('./pages/administratorOverview/administratorOverview'));
-app.use('/pl/administrator/dashboard', require('./pages/dashboard/dashboard'));
+app.use('/pl/administrator/dashboard', require('./pages/administratorDashboard/administratorDashboard'));
 
 //////////////////////////////////////////////////////////////////////
 //////////////////////////////////////////////////////////////////////
